Created main .cc
Yeah, it's a wall, I know... 😂 

Copypasted from ogdb, tested fairly extensively, but give it a whirl on the compiler and lmk.

Menu loop and switches for the twelve functions. Robust enough to handle input edge cases.

Left the functions as barebones as possible for now. This was mainly to get a working menu/switchboard/input loop working. The rest can be fleshed out as al ready started.

Will need some modification but for the most part it should be plug-and-play with this as a base.

Anyway, just tossing it up for your feedback etc. Lmk.